agentcore-neo4j-mcp-agent/pyproject.toml
[project]
name = "neo4j-mcp-agent"
version = "0.1.0"
description = "LangGraph ReAct agent for Neo4j MCP server deployed on AgentCore Runtime"
requires-python = ">=3.10"
dependencies = [
    "bedrock-agentcore",
    "strands-agents",
    "langgraph>=1.0.5",
    "langchain-mcp-adapters>=0.1.0",
    "langchain-aws>=0.2.0",
    "langchain-core>=0.3.0",
    "httpx>=0.27.0",
    "boto3>=1.35.0",
    "langchain>=0.3.0",
    "pyyaml>=6.0",
]
